created basic structure to reimport current hardening project
the current ssh hardening project doesn't hadhere to the styling and structure of modules which we require for this project. it must be created from the ground up to ensure all pieces are in place. the old configuration will now be imported on top

also note: unlike the default structure this one doesn't contain spec and tests. we want to unify everything under a common test framework for integration, so the relevance of these must be checked.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>