remove support for puppet 3.0.0
During rspec testing I noticed that puppet 3.0.0 wasn't able to include the parser functions as desired for saz/ssh. Since this is absolutely mandatory and puppet has since moved on, the decision to support puppet 3.0.0 will have to be revisited. At least for now, remove it in this module since it doesn't work with tests, while both stable branches of puppet work fine.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>